Split forgotten password view and action tests
